Second crack at migrations, migrations run now, but the first one does not work.

+++ cogs/triggered_responses.py
@@ -16,40 +16,48 @@ class TriggeredResponseCog(commands.Cog):
     @has_permissions(manage_messages=True)
     async def add_triggered_response(self, context, trigger: str, response: str):
         guild_id = context.guild.id
-        triggered_response, created = TriggeredResponse.get_or_create(
-            guild_id=guild_id,
-            trigger=trigger,
-            defaults={
-                'response': response
-            }
+        guild = Guild.get_or_none(
+            guild_id=guild_id
         )
-        if created:
-            print(f'{context.message.author.id} created a copy-pasta in '
-                  f'the guild: {guild_id} ({context.guild.name}). '
-                  f'Trigger: {trigger}, Response: {response}')
-        else:
-            triggered_response.response = response
-            triggered_response.save()
-            print(f'{context.message.author.id} updated the copy-pasta in '
-                  f'the guild: {guild_id} ({context.guild.name}). '
-                  f'Trigger: {trigger}, New Response: {response}')
-        await context.channel.send(response)
+        if guild:
+            triggered_response, created = TriggeredResponse.get_or_create(
+                guild=guild,
+                trigger=trigger,
+                defaults={
+                    'response': response
+                }
+            )
+            if created:
+                print(f'{context.message.author.id} created a copy-pasta in '
+                      f'the guild: {guild_id} ({context.guild.name}). '
+                      f'Trigger: {trigger}, Response: {response}')
+            else:
+                triggered_response.response = response
+                triggered_response.save()
+                print(f'{context.message.author.id} updated the copy-pasta in '
+                      f'the guild: {guild_id} ({context.guild.name}). '
+                      f'Trigger: {trigger}, New Response: {response}')
+            await context.channel.send(response)
 
     @commands.command(name='removepasta')
     @has_permissions(manage_messages=True)
     async def remove_triggered_response(self, context, trigger: str):
         guild_id = context.guild.id
-        triggered_response = TriggeredResponse.get_or_none(
-            guild_id=guild_id,
-            trigger=trigger
+        guild = Guild.get_or_none(
+            guild_id=guild_id
         )
-        if triggered_response is not None:
-            triggered_response.delete_instance()
-            await context.channel.send(f'Removed pasta "{trigger}" from the '
-                                       f'guild.')
-            print(f'{context.message.author.id} removed a copy-pasta in '
-                  f'the guild: {guild_id} ({context.guild.name}). '
-                  f'Trigger: {trigger}')
+        if guild:
+            triggered_response = TriggeredResponse.get_or_none(
+                guild=guild,
+                trigger=trigger
+            )
+            if triggered_response is not None:
+                triggered_response.delete_instance()
+                await context.channel.send(f'Removed pasta "{trigger}" from the '
+                                           f'guild.')
+                print(f'{context.message.author.id} removed a copy-pasta in '
+                      f'the guild: {guild_id} ({context.guild.name}). '
+                      f'Trigger: {trigger}')
 
     @commands.command(name='pastadelay')
     @has_permissions(manage_messages=True)
